src/requirements.txt
certifi==2020.4.5.1
flask-cors==3.0.10
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.13
urllib3==1.25.9
Werkzeug==1.0.1
uwsgi==2.0.19.1
fpdf2==2.3.5
Pillow==8.2.0
